requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.57.3
accelerate>=0.27.0
Pillow>=9.0.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
requests>=2.28.0
pycocoevalcap>=1.2      # for BLEU/CIDEr captioning metrics
pycocotools>=2.0.6      # COCO API
wandb>=0.15.0           # optional, for experiment tracking
